+++ src/etheroll.py
@@ -4,6 +4,7 @@ from os.path import expanduser
 
 from kivy.app import App
 from kivy.clock import Clock
+from kivy.logger import LOG_LEVELS, Logger
 from kivy.properties import NumericProperty, ObjectProperty, StringProperty
 from kivy.uix.boxlayout import BoxLayout
 from kivy.uix.floatlayout import FloatLayout
@@ -12,6 +13,9 @@ from kivy.utils import platform
 from kivymd.list import OneLineListItem
 from kivymd.theming import ThemeManager
 from kivymd.toolbar import Toolbar
+from raven import Client
+from raven.conf import setup_logging
+from raven.handlers.logging import SentryHandler
 
 from utils import Dialog, patch_find_library_android, patch_typing_python351
 from version import __version__
@@ -531,6 +535,53 @@ class Controller(FloatLayout):
             self.dialog_roll_success(tx_hash)
 
 
+class DebugRavenClient(object):
+    """
+    The DebugRavenClient should be used in debug mode, it just raises
+    the exception rather than capturing it.
+    """
+
+    def captureException(self):
+        raise
+
+
+def configure_sentry(in_debug=False):
+    """
+    Configure the Raven client, or create a dummy one if `in_debug` is `True`.
+    """
+    key = 'b290ecc8934f4cb599e6fa6af6cc5cc2'
+    # the public DSN URL is not available on the Python client
+    # so we're exposing the secret and will be revoking it on abuse
+    # https://github.com/getsentry/raven-python/issues/569
+    secret = '0ae02bcb5a75467d9b4431042bb98cb9'
+    project_id = '1111738'
+    dsn = 'https://{key}:{secret}@sentry.io/{project_id}'.format(
+        key=key, secret=secret, project_id=project_id)
+    if in_debug:
+        client = DebugRavenClient()
+    else:
+        client = Client(dsn=dsn, release=__version__)
+        # adds context for Android devices
+        if platform == 'android':
+            from jnius import autoclass
+            Build = autoclass("android.os.Build")
+            VERSION = autoclass('android.os.Build$VERSION')
+            android_os_build = {
+                'model': Build.MODEL,
+                'brand': Build.BRAND,
+                'device': Build.DEVICE,
+                'manufacturer': Build.MANUFACTURER,
+                'version_release': VERSION.RELEASE,
+            }
+            client.user_context({'android_os_build': android_os_build})
+        # Logger.error() to Sentry
+        # https://docs.sentry.io/clients/python/integrations/logging/
+        handler = SentryHandler(client)
+        handler.setLevel(LOG_LEVELS.get('error'))
+        setup_logging(handler)
+    return client
+
+
 class EtherollApp(App):
 
     theme_cls = ThemeManager()
@@ -539,5 +590,19 @@ class EtherollApp(App):
         self.icon = "docs/images/icon.png"
 
 
+def main():
+    # only send Android errors to Sentry
+    in_debug = platform != "android"
+    client = configure_sentry(in_debug)
+    try:
+        EtherollApp().run()
+    except Exception:
+        if type(client) == Client:
+            Logger.info(
+                'Errors will be sent to Sentry, run with "--debug" if you '
+                'are a developper and want to the error in the shell.')
+        client.captureException()
+
+
 if __name__ == '__main__':
-    EtherollApp().run()
+    main()