+++ src/main.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from etheroll import EtherollApp
+from etheroll import main
 
 if __name__ == '__main__':
-    EtherollApp().run()
+    main()